sanskara/requirements.txt
google-adk
fastapi
uvicorn
python-dotenv
loguru
pytest
pytest-asyncio
astrapy
websockets
google-search-results
redis
celery
pytest-asyncio
agentops
twilio